+++ packages/terra-site/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Converted component to use CSS modules
+* Updated Search Field example to demonstrate onInvalidSearch.
 
 1.1.0 - (July 18, 2017)
 ------------------